Handwritten_Text_generator/requirements.txt
torch
Pillow
argparse

